% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\name{demo_data}
\alias{demo_data}
\title{Demographic data}
\usage{
demo_data(
  seed = candidate_number,
  n = 150,
  age_m = 20.19,
  age_sd = 2.5,
  attr_m = 5.67,
  attr_sd = 1.34,
  d = 1,
  mark = FALSE,
  messy = TRUE
)
}
\description{
Demographic data
}
