% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/green_data.R
\name{green_data}
\alias{green_data}
\title{Green data}
\usage{
green_data(
  seed = candidate_number,
  n = 168,
  age_m = 20.19,
  age_sd = 2.5,
  es = 0.15,
  mark = FALSE,
  messy = TRUE
)
}
\description{
Simulates data based on Griskevicius, Tybur, & Van den Bergh (2010)
}
\details{
Griskevicius, V., Tybur, J. M., & Van den Bergh, B. (2010). Going green to be seen: status, reputation, and conspicuous conservation. Journal of personality and social psychology, 98(3), 392.
}
