% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_data.R
\name{red2021}
\alias{red2021}
\title{Datasets for AnD 2021 lab report}
\usage{
red2021(
  seed = candidate_number,
  n = 150,
  age_m = 20.19,
  age_sd = 2.5,
  attr_m = 5.67,
  attr_sd = 1.34,
  d = 1.5,
  mark = FALSE,
  messy = FALSE
)
}
\description{
Datasets for AnD 2021 lab report
}
